add DXLT v0.9.0 tests' outputs
The outputs of the DXLT tests shall be used as golden patterns in the
RTA unit / regression testing.

Only the outputs of the DXLT tests that have correspondent tests in RTA
are added.

The DXLT tests' outpus are not "as is"; they were changed to ease the
automatic checking / diff between RTA and DXLT.
The following updates were performed:
-renamed tests to match the RTA naming
-merged op_cipher_commands and op_mdha_commands in test_operation_alg
-all output was converted to lower case
-removed comments
-immediates, inlined data (keys etc.) modified to match the RTA output
(RTA endianess limitation)
-removed commands currently not supported by RTA (i.e. for SEC Era > 5)
-removed (seemingly) incorrect commands
-removed new commands in test_nfifo (not yet added in RTA)
-replaced commands in DXLT with the ones in RTA in cases where they have
the same effect; for e.g. in deco_dma_1, JUMP to offset 0xe4 (-0x1d) was
replaced with JUMP to 0: both commands perform the same thing - they jump
to the start of the job descriptor

Signed-off-by: Horia Geanta <horia.geanta@freescale.com>
Change-Id: I3b580887656450ba6b907b9cb6d53045da262051
Reviewed-on: http://git.am.freescale.net:8181/1004