fix SEQ IN PTR and SEQ OUT PTR for 36-bit pointers
Fix SEQ IN PTR and SEQ OUT PTR commands to handle correctly the cases
when pointers are wider than 32-bit (i.e. 36-bit, 40-bit etc.), i.e.
to generate a 2-word pointer field.

While here, change the type of the src / dst parameter of SEQ IN / OUT PTR
from uintptr_t to uint64_t.
In the SEQ IN PTR and SEQ OUT PTR, there would be no problem to use either
uint64_t or uintptr_t as the type of the parameter. This change is for
being consistent across all RTA commands, as follows: for FIFOLOAD, KEY,
LOAD, STORE uint64_t must be used and not uintptr_t. If uintptr_t is used,
we fail to cover the case when a 64-bit immediate value (not pointer!)
has to be inlined and the library is compiled on a 32-bit platform.
In this case, the immediate 64-bit value will be truncated to 32-bit
(since uintptr_t is unsigned int for 32-bit platforms).

Signed-off-by: Horia Geanta <horia.geanta@freescale.com>
Change-Id: I1a2149987f98cd32f19918b8c26545155fccda82
Reviewed-on: http://git.am.freescale.net:8181/1471
Tested-by: Stoica Cristian-B18196 <cristian.stoica@freescale.com>
Reviewed-by: Ambarus Tudor-Dan-B38632 <tudor.ambarus@freescale.com>
Reviewed-by: Stoica Cristian-B18196 <cristian.stoica@freescale.com>