fix anti-replay REPLAY error in WiMAX Descriptors
This error appears because the Share Type is NEVER and Descriptors
can execute in parallel (two or more DECOs process jobs using the same
Shared Descriptor at the same time). For WiMAX this translates into
multiple Encapsulation Output Frames with the same PN.

The solution consists in changing the Share Type to SERIAL (share
once the Descriptor has completed). This requires changing the way
of handling the key, the context information, the Input
and the Output Frame Pointers. Interest Job Descriptor fields
are written back to JD and not to the Shared Descriptor.

Signed-off-by: Tudor Ambarus <tudor.ambarus@freescale.com>
Change-Id: I42a0a20a790a269aab06d3eb4920bd06e14f7278
Reviewed-on: http://git.am.freescale.net:8181/2998
Reviewed-by: Geanta Neag Horia Ioan-B05471 <horia.geanta@freescale.com>
Tested-by: Geanta Neag Horia Ioan-B05471 <horia.geanta@freescale.com>