fix PDCP NULL int. decap. if out buf len is not a multiple of 8 bytes
If the final FIFO STORE in the NULL integrity descriptors
finishes at an alignment other than at the end of a dword, the
remainder of the last dword is popped and discarded. In this case,
the ICV (which should be 0x00 0x00 0x00 0x00), can no longer be
checked, and the descriptors ends with a watchdog timeout error
(0x4000_XX1C). This patch solves this issue by using the CONT
flag of the FIFO STORE command and thus the final dword is not
popped.
Introduced by commit 7852be6a00e3d1e94f0e7a19bf02265d1f623b15

Change-Id: I79ab6138f236b0bcdd2eda1d0c54907dfd97afeb
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/3258
Reviewed-by: Stoica Cristian-B18196 <cristian.stoica@freescale.com>
Reviewed-by: Geanta Neag Horia Ioan-B05471 <horia.geanta@freescale.com>
Tested-by: Geanta Neag Horia Ioan-B05471 <horia.geanta@freescale.com>