add DES-CBC/3DES-CBC & AES-XCBC-MAC-96 IPsec descriptors
These descriptors are supported only for platforms with
32-bit address pointers and SEC ERA 4 or higher.
The tunnel/transport mode of the IPsec ESP encapsulation descriptor
is supported only if the Outer/Transport IP Header is present
in the encapsulation output packet.
The tunnel/transport mode of the IPsec ESP decapsulation descriptor
is supported only if the Outer/Transport IP Header is present
in the decapsulation input packet.

The descriptor uses all the benefits of the built-in protocol
by computing the IPsec ESP with a hardware supported algorithms
combination (DES-CBC/3DES-CBC & HMAC-MD5-96).
The HMAC-MD5 authentication algorithm was chosen in order to
speed up the computational time for this intermediate step.
Warning: The user must allocate at least 32 bytes for
the authentication key (in order to use it also with HMAC-MD5-96),
even when using a shorter key for the AES-XCBC-MAC-96.

The encapsulation descriptor performs DES-CBC/3DES-CBC & HMAC-MD5-96
and then rereads the input packet to do the AES-XCBC-MAC-96 calculation
and to overwrite the MD5 ICV.
The decapsulation descriptor computes the AES-XCBC-MAC-96 to check
if the received ICV is correct, rereads the input packet to compute
the MD5 ICV, overwrites the XCBC ICV, and then sends the modified
input packet to the DES-CBC/3DES-CBC & HMAC-MD5-96 IPsec.

Change-Id: I9270b162761d66ee114886f87628d8f8c5ace348
Signed-off-by: Tudor Ambarus <tudor.ambarus@freescale.com>
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/4675
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Tested-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>