rename "DEBUG" flag to "RTA_DEBUG"
"DEBUG" flag in RTA guards the printing of error messages
in case the wrong parameters are passed to the functions
creating the descriptor instructions (i.e wrong parameters
for a certain SEC ERA). A conflict appears if the same
DEBUG flag is being used at the application layer to control
the verbosity of the application. Thus the flag in RTA has
been renamed to RTA_DEBUG in order to avoid the scenario
above.

Change-Id: I76f4603d6351d4da7b4dbb25602b4d4e2a8d0a12
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/5475
Reviewed-by: Geanta Neag Horia Ioan-B05471 <horia.geanta@freescale.com>
Tested-by: Geanta Neag Horia Ioan-B05471 <horia.geanta@freescale.com>