fix PDCP Control Plane with ZUC encryption & NULL integrity
The "direction" parameter sent to the ALG_OPERATION macro was
incorrectly set to OP_TYPE_ENCAP_PROTOCOL or OP_TYPE_DECAP_PROTOCOL
for PDCP Control Plane ZUC-E/NULL combination. This leads to
the operation to be skipped by RTA which leads to an error
when the user attempts to run the descriptor. This patch
modifies the corresponding parameter so the proper value
is being passed.

Change-Id: I6c42688201bbbc719dc8d1a3010605911d1b4c58
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/9623
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Tested-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>