MBMS processing should be done only on devices with SEC ERA >= 5
Initial requirement for MBMS processing was done for B4860 rev 1
and rev 2 (both having SEC ERA 5). There is no check that enforces
this, allowing the user to run the descriptors on lower SEC ERAs.
Some instructions are not available, leading to errors. This patch
ensures that the user cannot use MBMS processing on devices with
SEC ERA < 5.

Change-Id: Ibc0706196ad7d1b7875429d6211e7139c3411571
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/11379
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Tested-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>