WiMAX: invalidate prefetch buffer on encap (era5+)
This descriptor addreses the prefetch problem when modifying the
header of the input frame by invalidating the prefetch mechanism.

For performance reasons (due to the long read latencies),
the JQ will prefetch the input frame if a job cannot go immediately
into a DECO. As a result, the rewind is rewinding into the prefetch
buffer, not into memory. Therefore, in those cases where prefetch
is done, an unaware descriptor would update the memory but read
from the prefetched buffer and, as a result, it would not get
the updated header.

This descriptor invalidates the prefetch data and reads the updated
header from memory. The descriptor reads enough data to read to the
end of the prefetched data, dumps that data, rewinds the input frame
and just starts reading from the beginning again.

Change-Id: I48579c7e249033cff4cda01855d2be6be6b30970
Signed-off-by: Tudor Ambarus <tudor.ambarus@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/12819
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Tested-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>