fix PDCP Control Plane AES-CTR + SNOW f9 decapsulation
Due to the length of the descriptor, when the HFN override
mechanism is activated on P4080 R2 SoC, the length of the
descriptors exceeds the maximum allowable length. This patch
rewrites parts of the logic in the descriptor so that for
the aforementioned SoC the descriptor fits in the descriptor
buffer.

Change-Id: Ib44f38dbfd4744ca459b9609e971cd3c124cec5d
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/13519
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Tested-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>