fix WiFi for SEC ERA > 5
For SoCs with SEC ERA > 5 (i.e. T4240, T1040, T2080 etc.),
Errata A-005487 is fixed. But the patching of the loop where
the PN is reversed is done outside the if block guarding
the application of the workaround for SEC ERA <= 5. This leads
to a SEGFAULT for SEC ERA > 5.
This patch fixes the issue by moving the patch instruction
inside the block that is executed for SEC ERA <= 5

Change-Id: If5dd0945fa26a9371ed8dd9b98bbcdb97e7f80e7
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/13527
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Tested-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>