WiMAX: wait to update the header on encap era4
Sometimes the protocol operation started execution on the input frame
without waiting for the updated header to be written into memory.
This led to an invalid FCS value when FCS was activated.

This patch updated the encap descriptor in order to be sure
that the header is written in memory before starting the protocol
operation command.

Change-Id: If5dd45b3895aeddc720b4cb569bd37ac747cef93
Signed-off-by: Tudor Ambarus <tudor.ambarus@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/13562
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Tested-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>