add ICV checking operation for HMAC descriptor
This patch enables the user to request a descriptor for
performing HMAC ICV check.
While here, also remove the uneeded update of the VSOL
register (the SEQSTORE command will properly update
the SOL register). Also rename the "icv" parameter
and change its type to match the intended use in the
descriptor code.

Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/14650
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Tested-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Change-Id: I35ba580e6bc9a4137ac05678cf06464077d2e81a
Reviewed-on: http://git.am.freescale.net:8181/22733