add BIT macro for defining single-bit fields
BIT macro increases the readability of single-bit field definitions.

Though superfluous, *_SHIFT and *_MASK defines for single-bit fields
are not removed, in order to keep desc.h as close as possible to
desc.h from DXLT.

The macro is placed in compat.h, since it's already defined in
Linux kernel.

Change-Id: I63efa6d67e7b2d19ea9d2bef427aecd98bf01a92
Signed-off-by: Horia Geanta <horia.geanta@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/14748
Reviewed-by: Tudor-Dan Ambarus <tudor.ambarus@freescale.com>