IPsec new mode - add shared descriptors lengths
Add defines for IPsec encap / decap shared descriptors lengths.
These can be used at upper layers for deciding whether there is enough
space in the descriptor to inline PDB and/or keys.

While here, a more generic functionality is added.
rta_inline_query() takes the lengths of data items to be inlined in a
shared descriptor and returns a bitmask that tells the upper layer
what can be inlined and what will have to be accessed by reference.

Change-Id: Ie2264b5799540085f780f1ed5456e33533b89d65
Signed-off-by: Horia Geanta <horia.geanta@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/15435
Reviewed-by: Alexandru Porosanu <alexandru.porosanu@freescale.com>