IPsec new mode - add shared desc. lengths for NULL enc / auth
In the case of NULL encryption / authentication, the shared descriptors
base lengths are different than in the non-NULL case. That's due to
the fact that one KEY command is no longer needed in the descriptor.

Note that both encryption and authentication algorithms cannot be NULL
in the same time, according to standards.

Signed-off-by: Horia Geanta <horia.geanta@freescale.com>
Change-Id: If6f9f209b61aa1f173120e42004220ae9a17ebf1
Reviewed-on: http://git.am.freescale.net:8181/21760
Reviewed-by: Alexandru Porosanu <alexandru.porosanu@freescale.com>