fix MBMS Type 0 PDU error path halt condition
When a header CRC mismatch is detected for MBMS Type 0 PDUs,
only the statistics needs to be updated, the input frame
is not supposed to be copied in the output buffer. Thus,
the descriptor needs to stop executing after the statistics
have been written to the external memory.
This patch corrects the halt condition such as the halt occurs
only after the statistics were written.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>